# packages/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Vexel Health Platform API
  version: 1.0.0
  description: |
    Official API contract for Vexel Health Platform.

    Production tenancy is resolved by request Host (TenantDomain mapping).
    Clients do NOT choose tenant via header in production.
    The `x-tenant-id` header is allowed only in local DEV when `TENANCY_DEV_HEADER_ENABLED=1`.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /lab/tests:
    post:
      summary: Create LAB test definition
      operationId: createLabTest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestRequest'
      responses:
        '201':
          description: LAB test created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabTestDefinition'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/DomainError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    get:
      summary: List LAB test definitions
      operationId: listLabTests
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: LAB test list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLabTestsResponse'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /lab/tests/{testId}:
    get:
      summary: Get LAB test definition by id
      operationId: getLabTestById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/LabTestIdPathParam'
      responses:
        '200':
          description: LAB test definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabTestDefinition'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /lab/tests/{testId}/parameters:
    post:
      summary: Add parameter to LAB test definition
      operationId: addLabTestParameter
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/LabTestIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddParameterRequest'
      responses:
        '201':
          description: LAB parameter created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabTestParameter'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    get:
      summary: List parameters for LAB test definition
      operationId: listLabTestParameters
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/LabTestIdPathParam'
      responses:
        '200':
          description: LAB parameters list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLabTestParametersResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantIdHeader:
      name: x-tenant-id
      in: header
      description: DEV ONLY. Ignored in production. For local testing when hostname is localhost.
      required: false
      schema:
        type: string
        format: uuid
...
    LabTestIdPathParam:
      name: testId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Structured validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
    DomainError:
      description: Structured domain rule error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DomainErrorEnvelope'
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundErrorEnvelope'
    UnexpectedError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnexpectedErrorEnvelope'
  schemas:
...
    LabTestDefinition:
      type: object
      required:
        - id
        - code
        - name
        - department
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        department:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LabTestParameter:
      type: object
      required:
        - id
        - testId
        - name
        - displayOrder
        - active
      properties:
        id:
          type: string
          format: uuid
        testId:
          type: string
          format: uuid
        name:
          type: string
        unit:
          type: string
          nullable: true
        refLow:
          type: number
          nullable: true
        refHigh:
          type: number
          nullable: true
        refText:
          type: string
          nullable: true
        displayOrder:
          type: integer
        active:
          type: boolean
...
    ListLabTestsResponse:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LabTestDefinition'
        total:
          type: integer
    ListLabTestParametersResponse:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LabTestParameter'
        total:
          type: integer
...
    CreateTestRequest:
      type: object
      additionalProperties: false
      required:
        - code
        - name
        - department
      properties:
        code:
          type: string
        name:
          type: string
        department:
          type: string
        active:
          type: boolean
    AddParameterRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
        unit:
          type: string
        refLow:
          type: number
        refHigh:
          type: number
        refText:
          type: string
        displayOrder:
          type: integer
        active:
          type: boolean
...
    ValidationErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [validation_error]
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
          required:
            - type
            - fields
      required:
        - error
    DomainErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [domain_error]
            code:
              type: string
            message:
              type: string
          required:
            - type
            - code
            - message
      required:
        - error
...
    NotFoundErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [not_found]
            message:
              type: string
          required:
            - type
            - message
      required:
        - error
    UnexpectedErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [unexpected_error]
            message:
              type: string
            correlationId:
              type: string
          required:
            - type
            - message
      required:
        - error